Fix erroneous SQLGlot dependency in `semantic` by modularizing its own `commands.py`
